fix : bad practices against postgres wiki

+++ src/main/java/com/app/folioman/portfolio/domain/HarvestLotTracker.java
@@ -22,11 +22,11 @@ public class HarvestLotTracker {
             return;
         }
 
-        BigDecimal quantity = txn.getUnits() != null ? BigDecimal.valueOf(txn.getUnits()) : BigDecimal.ZERO;
+        BigDecimal quantity = txn.getUnits() != null ? txn.getUnits() : BigDecimal.ZERO;
         if (quantity.compareTo(BigDecimal.ZERO) == 0) {
             return;
         }
-        BigDecimal nav = txn.getNav() != null ? BigDecimal.valueOf(txn.getNav()) : BigDecimal.ZERO;
+        BigDecimal nav = txn.getNav() != null ? txn.getNav() : BigDecimal.ZERO;
 
         if (txn.getAmount().compareTo(BigDecimal.ZERO) > 0) {
             buy(txn.getTransactionDate(), quantity, nav);
